+++ docs/.vuepress/config.js
@@ -121,7 +121,7 @@ module.exports = {
         {
           title: 'Getting Started',
           collapsable: false,
-          children: ['getting-started/setup-with-react', 'getting-started/concepts']
+          children: ['getting-started/setup-with-react', 'getting-started/setup-with-vue', 'getting-started/concepts']
         },
         {
           title: 'Documentation',