+++ lib/state.ts
@@ -108,7 +108,10 @@ export class State<ValueType = any> {
     return this;
   }
 
-  // this creates a watcher that will fire a callback then destroy itself after invoking
+  /**
+   * @public
+   * Create a watcher that will fire a callback then destroy itself after invoking
+   */
   public onNext(callback: (value: ValueType) => void) {
     if (!this.watchers) this.watchers = {};
     this.watchers['_on_next_'] = () => {
@@ -117,6 +120,10 @@ export class State<ValueType = any> {
     };
   }
 
+  /**
+   * @public
+   * Set a name for this State, required to persist
+   */
   public key(key: string): this {
     // if persist was attempted before, but no key was provided retry persist function
     if (!this.name && this.persistState) this.persist(key);
@@ -124,14 +131,21 @@ export class State<ValueType = any> {
     return this;
   }
 
-  public type(type: any): this {
+  /**
+   * @public
+   * Fix the type to one of 'String', 'Boolean', 'Array', 'Object' or 'Number'
+   */
+  public type(type: TypeString | TypeConstructor): this {
     const supportedConstructors = ['String', 'Boolean', 'Array', 'Object', 'Number'];
-    if (typeof type === 'function' && supportedConstructors.includes(type.name)) {
-      this.typeOfVal = type.name.toLowerCase();
-    }
+    if (typeof type === 'function' && supportedConstructors.includes(type.name)) this.typeOfVal = type.name.toLowerCase();
+    else if (typeof type === 'string' && supportedConstructors.map(i => i.toLowerCase()).includes(type)) this.typeOfVal = type.toLowerCase();
     return this;
   }
 
+  /**
+   * @public
+   * Watch state for changes, run callback on each change
+   */
   public watch(key: number | string, callback: (value: any) => void): this {
     if (!this.watchers) this.watchers = {};
 
@@ -142,15 +156,27 @@ export class State<ValueType = any> {
     return this;
   }
 
-  public undo() {
-    this.set(this.previousState);
-  }
-
+  /**
+   * @public
+   * Remove watcher by key
+   */
   public removeWatcher(key: number | string): this {
     delete this.watchers[key];
     return this;
   }
 
+  /**
+   * @public
+   * Restore previous state
+   */
+  public undo() {
+    this.set(this.previousState);
+  }
+
+  /**
+   * @public
+   * If State is boolean, invert
+   */
   public toggle(): this {
     if (typeof this._value === 'boolean') {
       // @ts-ignore
@@ -159,42 +185,75 @@ export class State<ValueType = any> {
     return this;
   }
 
+  /**
+   * @public
+   * Reset the State to as declared
+   */
   public reset(): this {
-    reset(this);
+    this.isSet = false;
+    this.previousState = null;
+    this.privateWrite(this.initialState);
+    if (this.persistState) this.instance().storage.remove(this.name);
     return this;
   }
 
-  // returns a fresh copy of the current value
+  /**
+   * @public
+   * Returns a copy of the current value, objects and arrays will be cloned
+   */
   public copy(): any {
     return copy(this.value);
   }
-
+  /**
+   * @public
+   * Is the value equal to parameter
+   */
   public is(x: any) {
     return this.value === x;
   }
 
+  /**
+   * @public
+   * Is the value not equal to parameter
+   */
   public isNot(x: any) {
     return this.value !== x;
   }
 
-  // INTERNAL
+  /**
+   * @internal
+   * Write value directly to State
+   */
   public privateWrite(value: any) {
     this._value = copy(value);
     this.nextState = copy(value);
+    // If
     if (this.persistState) this.instance().storage.set(this.name, this.getPersistableValue());
   }
 
+  /**
+   * @internal
+   *
+   */
   private isCorrectType(value): boolean {
     let type: string = typeof value;
     if (type === 'object' && Array.isArray(value)) type = 'array';
     return type === this.typeOfVal;
   }
 
+  /**
+   * @internal
+   *
+   */
   public destroy(): void {
     this.dep.deps.clear();
     this.dep.subs.clear();
   }
 
+  /**
+   * @internal
+   *
+   */
   public getPersistableValue(): any {
     return this.value;
   }
@@ -215,11 +274,6 @@ export const StateGroup = (instance: () => Pulse, stateGroup: Object): any => {
 
 export default State;
 
-export function reset(instance: State) {
-  instance.isSet = false;
-  instance.previousState = null;
-  instance.privateWrite(instance.initialState);
-  if (instance.persistState) instance.instance().storage.remove(instance.name);
-}
-
 export type SetFunc<ValueType> = (state: ValueType) => ValueType;
+type TypeString = 'string' | 'boolean' | 'array' | 'object' | 'number';
+type TypeConstructor = StringConstructor | BooleanConstructor | ArrayConstructor | ObjectConstructor | NumberConstructor;